Move assembly files into submodules
